Fix [Batch Run] Number input for int and float parameters is malfunctioning.
